Broken commit: tons-o-changes...
Bumped version so once we get this functional we'll be at 0.0.1. Also
updated the gemspec; we now officially rely on the meta_where gem.

Added and "integrated" the model helper extensions (this part is really
not done yet). The main part of this is the addition of the
Axis::Attribute class and its two, nested helper classes, Filter and
Sort.

TODO: Update lib/axis/model.rb to correctly use lib/axis/attribute.rb