Added manual rails extension file.
This new file allows apps that might not run in the context of a full
rails environment, yet that want to extend the core rails classes to
make the various helper methods available on them, to "require" this
file to manually force extensions to be applied:

require 'axis/enable'

Normal rails apps will NEVER need to do this. This is primarily
available for some forms of testing.

This is all in preparation for adding a railtie to do this extension
when the rails environment loads. When we do this we'll modify the main
"axis.rb" file so it doesn't manually extend rails types.